Intcode updates following later day

+++ 2019/rust/day_05/src/part_1.rs
@@ -1,6 +1,6 @@
 use intcode::{IntcodeState, run_intcode};
 
-pub fn solve(input: &str) -> i32 {
+pub fn solve(input: &str) -> i64 {
     let intcode = IntcodeState::new(input, vec![1]);
 
     let mut result = run_intcode(intcode);
